Add files via upload
This is an RMarkdown .html file which covers Experiment 1 of the following publication: https://moralitylab.bc.edu/wp-content/uploads/2022/01/mcmanus2021final.pdf. This file not only contains code and output that transforms raw data into publication- or deliverable-ready statistics and visualizations. What makes this file special is that it is neatly organized into a readable .html file (and therefore can be the deliverable itself) to be sent to teammates or coworkers.